Create labjack_analog_broadcaster.py
A first attempt at hacking the labjack example code for CORIOLIX.  Essentially, I took the "take 10 readings from the labjack" part of the example and made it into a while loop so it kept on reading in data.  I then added the socket module to the imports and created a socket connection to local network on port 30325.  Each analog sample is sent out on that socket.